Refactor team agent functional tests
